Split integration tests for OLS and debug endpoints
